+++ src/main/java/com/example/forumproject/models/Post.java
@@ -11,7 +11,7 @@ public class Post {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "user_id")
@@ -39,7 +39,7 @@ public class Post {
     public Post() {
     }
 
-    public Post(int id,
+    public Post(Long id,
                 User author,
                 String title,
                 String content) {
@@ -49,11 +49,11 @@ public class Post {
         this.content = content;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
